+++ packages/desktop/src/renderer/Root.tsx
@@ -22,6 +22,7 @@ import LoadingPanel from './components/LoadingPanel/LoadingPanel'
 import { HTML5Backend } from 'react-dnd-html5-backend'
 import { DndProvider } from 'react-dnd'
 import { ErrorModal } from './components/ui/ErrorModal/ErrorModal'
+import SearchModal from './components/SearchModal/SearchModal'
 
 export const persistor = persistStore(store)
 export default () => {
@@ -33,6 +34,7 @@ export default () => {
             <Provider store={store}>
               <PersistGate loading={null} persistor={persistor}>
                 <SentryWarning />
+                <SearchModal />
                 <ErrorModal />
                 <LoadingPanel />
                 <CreateChannel />