+++ packages/desktop/src/renderer/components/widgets/settings/SettingsModal/SettingsModal.test.tsx
@@ -22,7 +22,7 @@ describe('SettingsModal', () => {
           aria-hidden="true"
         />
         <div
-          class="Modalroot MuiModal-root css-1hk9oxe-MuiModal-root"
+          class="Modalroot MuiModal-root css-4kftk5-MuiModal-root"
           role="presentation"
         >
           <div
@@ -521,7 +521,7 @@ describe('SettingsModal', () => {
           aria-hidden="true"
         />
         <div
-          class="Modalroot MuiModal-root css-1hk9oxe-MuiModal-root"
+          class="Modalroot MuiModal-root css-4kftk5-MuiModal-root"
           role="presentation"
         >
           <div