+++ packages/mobile/src/components/CreateCommunity/CreateCommunity.types.ts
@@ -1,0 +1,4 @@
+export interface CreateCommunityProps {
+  createCommunityAction: (name: string) => void
+  redirectionAction: () => void
+}