+++ packages/mobile/storybook/index.js
@@ -3,9 +3,10 @@ import './rn-addons'
 import { LogBox } from 'react-native'
 
 import { withKnobs } from '@storybook/addon-knobs'
+
 import {
-  addDecorator,
   configure,
+  addDecorator,
   getStorybookUI,
 } from '@storybook/react-native'
 
@@ -18,12 +19,14 @@ addDecorator(withNavigation)
 
 configure(() => {
   require('../src/components/JoinCommunity/JoinCommunity.stories')
+  require('../src/components/CreateCommunity/CreateCommunity.stories')
   require('../src/components/Appbar/Appbar.stories')
   require('../src/components/Registration/UsernameRegistration.stories')
   require('../src/components/ChannelTile/ChannelTile.stories')
   require('../src/components/ChannelList/ChannelList.stories')
   require('../src/components/Message/Message.stories')
   require('../src/components/Chat/Chat.stories')
+  require('../src/components/TextWithLink/TextWithLink.stories')
   require('../src/components/Typography/Typography.stories')
   require('../src/components/Button/Button.stories')
   require('../src/components/Input/Input.stories')